Internal

LocalSearchSolvers.AbstractSolver — Type

AbstractSolver

Abstract type to encapsulate the different solver types such as Solver or _SubSolver.

LocalSearchSolvers.Constraint — Type

Constraint{F <: Function}

Structure to store an error function and the variables it constrains.

LocalSearchSolvers.MOIAllDifferent — Type

MOIAllDifferent <: MOI.AbstractVectorSet

DOCSTRING

LocalSearchSolvers.MOIAllEqual — Type

MOIAllEqual <: MOI.AbstractVectorSet

DOCSTRING

LocalSearchSolvers.MOIAllEqualParam — Type

MOIAllEqualParam{T <: Number} <: MOI.AbstractVectorSet

DOCSTRING

Arguments:

param::T: DESCRIPTION

dimension::Int: DESCRIPTION

MOIAllEqualParam(param, dim = 0) = begin #= none:5 =# new{typeof(param)}(param, dim) end: DESCRIPTION

LocalSearchSolvers.MOIAlwaysTrue — Type

MOIAlwaysTrue <: MOI.AbstractVectorSet

DOCSTRING

LocalSearchSolvers.MOIDistDifferent — Type

MOIDistDifferent <: MOI.AbstractVectorSet

DOCSTRING

LocalSearchSolvers.MOIEq — Type

MOIEq <: MOI.AbstractVectorSet

DOCSTRING

LocalSearchSolvers.MOIError — Type

MOIError{F <: Function} <: MOI.AbstractVectorSet

DOCSTRING

Arguments:

f::F: DESCRIPTION

dimension::Int: DESCRIPTION

MOIError(f, dim = 0) = begin #= none:5 =# new{typeof(f)}(f, dim) end: DESCRIPTION

LocalSearchSolvers.MOIOrdered — Type

MOIOrdered <: MOI.AbstractVectorSet

DOCSTRING

LocalSearchSolvers.MOIPredicate — Type

MOIPredicate{F <: Function} <: MOI.AbstractVectorSet

DOCSTRING

Arguments:

f::F: DESCRIPTION

dimension::Int: DESCRIPTION

MOIPredicate(f, dim = 0) = begin #= none:5 =# new{typeof(f)}(f, dim) end: DESCRIPTION

LocalSearchSolvers.Objective — Type

Objective{F <: Function}

A structure to handle objectives in a solver. struct Objective{F <: Function} name::String f::F end`

LocalSearchSolvers.Objective — Method

Objective(F, o::Objective{F2}) where {F2 <: Function}

Constructor used in specializing a solver. Should never be called externally.

LocalSearchSolvers.Optimizer — Type

Optimizer <: MOI.AbstractOptimizer

DOCSTRING

Arguments:

solver::Solver: DESCRIPTION

status::MOI.TerminationStatusCode: DESCRIPTION

options::Options: DESCRIPTION

LocalSearchSolvers.Optimizer — Type

Optimizer(model = Model(); options = Options())

DOCSTRING

LocalSearchSolvers.SudokuInstance — Type

mutable struct SudokuInstance{T <: Integer} <: AbstractMatrix{T}

A struct for SudokuInstances, which is a subtype of AbstractMatrix.

SudokuInstance(A::AbstractMatrix{T})
SudokuInstance(::Type{T}, n::Int) # fill in blank sudoku of type T
SudokuInstance(n::Int) # fill in blank sudoku of type Int
SudokuInstance(::Type{T}) # fill in "standard" 9×9 sudoku of type T
SudokuInstance() # fill in "standard" 9×9 sudoku of type Int
SudokuInstance(n::Int, P::Pair{Tuple{Int, Int}, T}...) where {T <: Integer} # construct a sudoku given pairs of coordinates and values
SudokuInstance(P::Pair{Tuple{Int, Int}, T}...) # again, default to 9×9 sudoku, constructing given pairs

Constructor functions for the SudokuInstance struct.

LocalSearchSolvers.SudokuInstance — Method

SudokuInstance(X::Dictionary)

Construct a SudokuInstance with the values X of a solver as input.

LocalSearchSolvers.Variable — Type

Variable{D <: AbstractDomain}

A structure containing the necessary information for a solver's variables: name, domain, and constraints it belongs.

struct Variable{D <: AbstractDomain}
    domain::D
    constraints::Indices{Int}
end

LocalSearchSolvers._Model — Type

_Model{V <: Variable{<:AbstractDomain},C <: Constraint{<:Function},O <: Objective{<:Function}}

A struct to model a problem as a set of variables, domains, constraints, and objectives.

struct _Model{V <: Variable{<:AbstractDomain},C <: Constraint{<:Function},O <: Objective{<:Function}}
    variables::Dictionary{Int,V}
    constraints::Dictionary{Int,C}
    objectives::Dictionary{Int,O}

    # counter to add new variables: vars, cons, objs
    max_vars::Ref{Int}
    max_cons::Ref{Int}
    max_objs::Ref{Int}

    # Bool to indicate if the _Model instance has been specialized (relatively to types)
    specialized::Ref{Bool}

    # Symbol to indicate the kind of model for specialized methods such as pretty printing
    kind::Symbol
end

LocalSearchSolvers._State — Type

_State{T <: Number}

A mutable structure to store the internal state of a solver. All methods applied to _State are forwarded to S <: AbstractSolver.

mutable struct _State{T <: Number}
    values::Dictionary{Int,T} # TODO: handle multiple value type
    vars_costs::Dictionary{Int,Float64}
    cons_costs::Dictionary{Int,Float64}
    error::Float64
    tabu::Dictionary{Int,Int}
    optimizing::Bool
    best_solution::Dictionary{Int,T}
    best_solution_value::Union{Nothing,T}
end

LocalSearchSolvers._SubSolver — Type

_SubSolver <: AbstractSolver

An internal solver type called by Solver when multithreading is enabled.

Arguments:

id::Int: subsolver id for debugging

model::Model: a ref to the model of the main solver

state::_State: a deepcopy of the main solver that evolves independently

options::Options: a ref to the options of the main solver

LocalSearchSolvers._SubSolver — Method

_SubSolver(ms::Solver, id)

Internal structure used in multithreading and distributed version of the solvers. It is only created at the start of a solve! run. Its behaviour regarding to sharing information is determined by the main Solver.

Base.Multimedia.display — Method

display(io::IO, S::SudokuInstance)
display(S::SudokuInstance) # default to stdout

Displays an $n\times n$ SudokuInstance.

Base.Multimedia.display — Method

Base.display(X, Val(:sudoku))

Extends Base.display to a sudoku configuration.

Base.Multimedia.display — Method

Base.display(X::Dictionary)

Extends Base.display to a sudoku configuration.

Base.Multimedia.display — Method

Base.display(S::SudokuInstance)

Extends Base.display to SudokuInstance.

Base.copy — Method

Base.copy(set::DiscreteSet) = begin

DOCSTRING

Base.copy — Method

Base.copy(set::MOIError) = begin

DOCSTRING

Base.empty! — Method

empty!(m::Model)

DOCSTRING

Base.empty! — Method

empty!(s::_State)

DOCSTRING

Base.empty! — Method

empty!(s::Solver)

DOCSTRING

Base.getindex — Method

Base.getindex(S::SudokuInstance, i::Int)
Base.getindex(S::SudokuInstance, I::Vararg{Int,N}) where {N}

Extends Base.getindex for SudokuInstance.

Base.setindex! — Method

Base.setindex!(S::SudokuInstance, v, i::Int)
Base.setindex!(S::SudokuInstance, v, I::Vararg{Int,N})

Extends Base.setindex! for SudokuInstance.

Base.size — Method

Base.size(S::SudokuInstance)

Extends Base.size for SudokuInstance.

ConstraintDomains._add! — Method

_add!(c::Constraint, x)

Add the variable of indice x to c.

ConstraintDomains._delete! — Method

_delete!(c::Constraint, x::Int)

Delete x from c.

ConstraintDomains._get_domain — Method

_get_domain(x::Variable) = begin

DOCSTRING

ConstraintDomains._length — Method

_length(c::Constraint)

Return the number of constrained variables by c.

JuMP.build_variable — Method

JuMP.build_variable(::Function, info::JuMP.VariableInfo, set::T) where T <: MOI.AbstractScalarSet

DOCSTRING

Arguments:

``: DESCRIPTION

info: DESCRIPTION

set: DESCRIPTION

LocalSearchSolvers._add_to_constraint! — Method

_add_to_constraint!(x::Variable, id)

Add a constraint id to the list of contraints of x.

LocalSearchSolvers._best! — Method

_best!(s::S, val, values = Dictionary()) where S <: Union{_State, AbstractSolver}

Set the best known value to val and, if values not empty, the best known solution.

LocalSearchSolvers._best — Method

_best(s::S) where S <: Union{_State, AbstractSolver}

Access the best known solution value (defined for optimization models only).

LocalSearchSolvers._check_restart — Method

_check_restart(s)

Check if a restart of s is necessary. If s has subsolvers, this check is independent for all of them.

LocalSearchSolvers._check_subs — Method

_check_subs(s)

Check if any subsolver of a main solver s, for

Satisfaction, has a solution, then return it, resume the run otherwise

Optimization, has a better solution, then assign it to its internal state

LocalSearchSolvers._compute! — Method

_compute!(s; o::Int = 1, cons_lst = Indices{Int}())

Compute the objective o's value if s is satisfied and return the current error.

Arguments:

s: a solver

o: targeted objective

cons_lst: list of targeted constraints, if empty compute for the whole set

LocalSearchSolvers._compute_cost! — Method

_compute_cost!(s, ind, c)

Compute the cost of constraint c with index ind.

LocalSearchSolvers._compute_costs! — Method

_compute_costs!(s; cons_lst::Indices{Int} = Indices{Int}())

Compute the cost of constraints c in cons_lst. If cons_lst is empty, compute the cost for all the constraints in s.

LocalSearchSolvers._compute_objective! — Method

_compute_objective!(s, o::Objective)
_compute_objective!(s, o = 1)

Compute the objective o's value.

LocalSearchSolvers._cons_cost! — Method

_cons_cost!(s::S, c, cost) where S <: Union{_State, AbstractSolver}

Set the cost of constraint c.

LocalSearchSolvers._cons_cost — Method

_cons_cost(s::S, c) where S <: Union{_State, AbstractSolver}

Return the cost of constraint c.

LocalSearchSolvers._cons_costs! — Method

_cons_costs!(s::S, costs) where S <: Union{_State, AbstractSolver}

Set the constraints costs.

LocalSearchSolvers._cons_costs — Method

_cons_costs(s::S) where S <: Union{_State, AbstractSolver}

Access the constraints costs.

LocalSearchSolvers._constriction — Method

_constriction(x::Variable)

Return the cosntriction of x, i.e. the number of constraints restricting x.

LocalSearchSolvers._decay_tabu! — Method

_decay_tabu!(s::S) where S <: Union{_State, AbstractSolver}

Decay the tabu list.

LocalSearchSolvers._decrease_tabu! — Method

_decrease_tabu!(s::S, x) where S <: Union{_State, AbstractSolver}

Decrement the tabu value of variable x.

LocalSearchSolvers._delete_from_constraint! — Method

_delete_from_constraint!(x::Variable, id)

Delete a constraint id from the list of contraints of x.

LocalSearchSolvers._delete_tabu! — Method

_delete_tabu!(s::S, x) where S <: Union{_State, AbstractSolver}

Delete the tabu entry of variable x.

LocalSearchSolvers._draw! — Method

_draw!(s)

Draw a random (re-)starting configuration.

LocalSearchSolvers._dynamic! — Method

_dynamic!(options, dynamic) = begin

DOCSTRING

LocalSearchSolvers._dynamic — Method

_dynamic(options) = begin

DOCSTRING

LocalSearchSolvers._empty_tabu! — Method

_empty_tabu!(s::S) where S <: Union{_State, AbstractSolver}

Empty the tabu list.

LocalSearchSolvers._error! — Method

_error!(s::S, val) where S <: Union{_State, AbstractSolver}

Set the error of the current state of s to val.

LocalSearchSolvers._error — Method

_error(s::S) where S <: Union{_State, AbstractSolver}

Access the error of the current state of s.

LocalSearchSolvers._find_rand_argmax — Method

_find_rand_argmax(d::DictionaryView)

Compute argmax of d and select one element randomly.

LocalSearchSolvers._format_line — Method

_format_line(r, M)

Format line of a sudoku grid.

LocalSearchSolvers._format_line_segment — Method

_format_line_segment(r, col_pos, M)

Format line segment of a sudoku grid.

LocalSearchSolvers._format_val — Method

_format_val(a)

Format an integer a into a string for SudokuInstance.

LocalSearchSolvers._get_constraints — Method

_get_constraints(x::Variable)

Access the list of constraints of x.

LocalSearchSolvers._get_sep_line — Method

_get_sep_line(s, pos_row, M)

Return a line separator.

LocalSearchSolvers._get_vars — Method

_get_vars(c::Constraint)

Returns the variables constrained by c.

LocalSearchSolvers._inc_cons! — Method

_inc_vars!(m::M) where M <: Union{Model, AbstractSolver}

Increment the maximum constraint id that has been attributed to m.

LocalSearchSolvers._inc_objs! — Method

_inc_vars!(m::M) where M <: Union{Model, AbstractSolver}

Increment the maximum objective id that has been attributed to m.

LocalSearchSolvers._inc_vars! — Method

_inc_vars!(m::M) where M <: Union{Model, AbstractSolver}

Increment the maximum variable id that has been attributed to m.

LocalSearchSolvers._init_solve! — Method

_init_solve!(s::S) where S <: AbstractSolver

Initialize a solver in both sequential and parallel contexts.

LocalSearchSolvers._insert_tabu! — Method

_insert_tabu!(s::S, x, tabu_time) where S <: Union{_State, AbstractSolver}

Insert the bariable x as tabu for tabu_time.

LocalSearchSolvers._is_empty — Method

_is_empty(m::Model)

DOCSTRING

LocalSearchSolvers._iteration! — Method

_iteration!(options, iterations) = begin

DOCSTRING

LocalSearchSolvers._iteration — Method

_iteration(options) = begin

DOCSTRING

LocalSearchSolvers._length_tabu — Method

_length_tabu!(s::S) where S <: Union{_State, AbstractSolver}

Return the length of the tabu list.

LocalSearchSolvers._max_cons — Method

_max_cons(m::M) where M <: Union{Model, AbstractSolver}

Access the maximum constraint id that has been attributed to m.

LocalSearchSolvers._max_objs — Method

_max_objs(m::M) where M <: Union{Model, AbstractSolver}

Access the maximum objective id that has been attributed to m.

LocalSearchSolvers._max_vars — Method

_max_vars(m::M) where M <: Union{Model, AbstractSolver}

Access the maximum variable id that has been attributed to m.

LocalSearchSolvers._move! — Function

_move!(s, x::Int, dim::Int = 0)

Perform an improving move in x neighbourhood if possible.

Arguments:

s: a solver of type S <: AbstractSolver

x: selected variable id

dim: describe the dimension of the considered neighbourhood

LocalSearchSolvers._neighbours — Function

_neighbours(s, x, dim = 0)

DOCSTRING

Arguments:

s: DESCRIPTION

x: DESCRIPTION

dim: DESCRIPTION

LocalSearchSolvers._optimizing! — Method

_optimizing!(s::S) where S <: Union{_State, AbstractSolver}

Set the solver optimizing status to true.

LocalSearchSolvers._optimizing — Method

_optimizing(s::S) where S <: Union{_State, AbstractSolver}

Check if s is in an optimizing state.

LocalSearchSolvers._print_level! — Method

_print_level!(options, level) = begin

DOCSTRING

LocalSearchSolvers._print_level — Method

_print_level(options) = begin

DOCSTRING

LocalSearchSolvers._restart! — Function

_restart!(s, k = 10)

Restart a solver.

LocalSearchSolvers._satisfying! — Method

_satisfying!(s::S) where S <: Union{_State, AbstractSolver}

Set the solver optimizing status to false.

LocalSearchSolvers._select_worse — Method

_select_worse(s::S) where S <: Union{_State, AbstractSolver}

Within the non-tabu variables, select the one with the worse error .

LocalSearchSolvers._set! — Method

_set!(s::S, x, val) where S <: Union{_State, AbstractSolver}

Set the value of variable x to val.

LocalSearchSolvers._set_domain! — Method

_set_domain!(m::Model, x, values)

DOCSTRING

Arguments:

m: DESCRIPTION

x: DESCRIPTION

values: DESCRIPTION

LocalSearchSolvers._solution! — Method

_solution!(s::S, values) where S <: Union{_State, AbstractSolver}

Set the best known solution to values.

LocalSearchSolvers._solution — Method

_solution(s::S) where S <: Union{_State, AbstractSolver}

Access the best known solution.

LocalSearchSolvers._solutions! — Method

_solutions!(options, solutions) = begin

DOCSTRING

LocalSearchSolvers._solutions — Method

_solutions(options) = begin

DOCSTRING

LocalSearchSolvers._solve! — Method

_solve!(s, stop)

Start a solving run of a subsolver with a shared Atomic boolean stop.

LocalSearchSolvers._specialize! — Method

_specialize!(options, specialize) = begin

DOCSTRING

LocalSearchSolvers._specialize — Method

_specialize(options) = begin

DOCSTRING

LocalSearchSolvers._step! — Method

_step!(s)

Iterate a step of the solver run.

LocalSearchSolvers._swap_value! — Method

_set!(s::S, x, y) where S <: Union{_State, AbstractSolver}

Swap the values of variables x and y.

LocalSearchSolvers._tabu! — Method

_tabu!(s::S, tabu) where S <: Union{_State, AbstractSolver}

Set the variables tabu list.

LocalSearchSolvers._tabu — Method

_tabu(s::S, x) where S <: Union{_State, AbstractSolver}

Return the tabu value of variable x.

LocalSearchSolvers._tabu — Method

_tabu(s::S) where S <: Union{_State, AbstractSolver}

Access the list of tabu variables.

LocalSearchSolvers._tabu_delta! — Method

_tabu_delta!(options, time) = begin

DOCSTRING

LocalSearchSolvers._tabu_delta — Method

_tabu_delta(options) = begin

DOCSTRING

LocalSearchSolvers._tabu_local! — Method

_tabu_local!(options, time) = begin

DOCSTRING

LocalSearchSolvers._tabu_local — Method

_tabu_local(options) = begin

DOCSTRING

LocalSearchSolvers._tabu_time! — Method

_tabu_time!(options, time) = begin

DOCSTRING

LocalSearchSolvers._tabu_time — Method

_tabu_time(options) = begin

DOCSTRING

LocalSearchSolvers._threads! — Method

_threads!(options, threads) = begin

DOCSTRING

LocalSearchSolvers._threads — Method

_threads(options) = begin

DOCSTRING

LocalSearchSolvers._time_limit! — Method

_time_limit!(options, time::Time) = begin

DOCSTRING

LocalSearchSolvers._time_limit — Method

_time_limit(options) = begin

DOCSTRING

LocalSearchSolvers._to_union — Method

_to_union(datatype)

Make a minimal Union type from a collection of data types.

LocalSearchSolvers._value! — Method

_value!(s::S, x, val) where S <: Union{_State, AbstractSolver}

Set the value of variable x to val.

LocalSearchSolvers._value — Method

_value(s::S, x) where S <: Union{_State, AbstractSolver}

Return the value of variable x.

LocalSearchSolvers._values! — Method

_values!(s::S, values) where S <: Union{_State, AbstractSolver}

Set the variables values.

LocalSearchSolvers._values — Method

_vars_costs(s::S) where S <: Union{_State, AbstractSolver}

Access the variables costs.

LocalSearchSolvers._var_cost! — Method

_var_cost!(s::S, x, cost) where S <: Union{_State, AbstractSolver}

Set the cost of variable x.

LocalSearchSolvers._var_cost — Method

_var_cost(s::S, x) where S <: Union{_State, AbstractSolver}

Return the cost of variable x.

LocalSearchSolvers._vars_costs! — Method

_vars_costs!(s::S, costs) where S <: Union{_State, AbstractSolver}

Set the variables costs.

LocalSearchSolvers._vars_costs — Method

_vars_costs(s::S) where S <: Union{_State, AbstractSolver}

Access the variables costs.

LocalSearchSolvers._verbose — Method

_verbose(settings, str)

Temporary logging function. #TODO: use better log instead (LoggingExtra.jl)

LocalSearchSolvers.domain_size — Method

domain_size(m::Model, x) = begin

DOCSTRING

LocalSearchSolvers.get_kind — Method

get_kind(m::M) where M <: Union{Model, AbstractSolver}

Access the kind of m, such as :sudoku or :generic (default).

LocalSearchSolvers.is_sat — Method

is_sat(m::M) where M <: Union{Model, AbstractSolver}

Return true if m is a satisfaction model.

LocalSearchSolvers.is_specialized — Method

is_specialized(m::M) where M <: Union{Model, AbstractSolver}

Return true if the model is already specialized.

LocalSearchSolvers.length_objs — Method

length_objs(m::M) where M <: Union{Model, AbstractSolver}

Return the number of objectives in m.

LocalSearchSolvers.length_vars — Method

length_vars(m::M) where M <: Union{Model, AbstractSolver}

Return the number of variables in m.

LocalSearchSolvers.max_domains_size — Method

max_domains_size(m::Model, vars) = begin

DOCSTRING

LocalSearchSolvers.set_status! — Method

set_status!(optimizer::Optimizer, status::Symbol)

DOCSTRING

MathOptInterface.Utilities.supports_default_copy_to — Method

Copy constructor for the optimizer

MathOptInterface.add_constraint — Method

MOI.add_constraint(optimizer::Optimizer, vars::MOI.VectorOfVariables, set::MOIError)

DOCSTRING

Arguments:

optimizer: DESCRIPTION

vars: DESCRIPTION

set: DESCRIPTION

MathOptInterface.add_constraint — Method

MOI.add_constraint(optimizer::Optimizer, v::SVF, set::DiscreteSet{T}) where T <: Number

DOCSTRING

Arguments:

optimizer: DESCRIPTION

v: DESCRIPTION

set: DESCRIPTION

MathOptInterface.add_variable — Method

MOI.add_variable(model::Optimizer) = begin

DOCSTRING

MathOptInterface.empty! — Method

MOI.empty!(opt) = begin

DOCSTRING

MathOptInterface.get — Method

MOI.get(::Optimizer, ::MOI.SolverName) = begin

DOCSTRING

MathOptInterface.is_empty — Method

MOI.is_empty(model::Optimizer) = begin

DOCSTRING

MathOptInterface.optimize! — Method

MOI.optimize!(model::Optimizer)

MathOptInterface.set — Function

MOI.set(::Optimizer, ::MOI.Silent, bool = true) = begin

DOCSTRING

Arguments:

``: DESCRIPTION

``: DESCRIPTION

bool: DESCRIPTION

MathOptInterface.supports_constraint — Method

MOI.supports_constraint(::Optimizer, ::Type{VOV}, ::Type{MOIError}) = begin

DOCSTRING

Arguments:

``: DESCRIPTION

``: DESCRIPTION

``: DESCRIPTION