Public

Public

LocalSearchSolvers.DiscreteSet — Type

DiscreteSet(values)

LocalSearchSolvers.Options — Type

Options

DOCSTRING

Arguments:

dynamic::Bool: DESCRIPTION

iteration::Union{Int, Float64}: DESCRIPTION

print_level::Symbol: DESCRIPTION

solutions::Int: DESCRIPTION

specialize::Bool: DESCRIPTION

tabu_time::Int: DESCRIPTION

tabu_local::Int: DESCRIPTION

tabu_delta::Float64: DESCRIPTION

threads::Int: DESCRIPTION

time_limit::Float64: DESCRIPTION

`function Options(; dynamic = false, iteration = 1000, printlevel = :minimal, solutions = 1, specialize = !dynamic, tabutime = 0, tabulocal = 0, tabudelta = 0.0, threads = typemax(0), timelimit = Time(0)) #= none:13 =# #= none:25 =# dsstr = "The model types are specialized to the starting domains, constraints," * " and objectives types. Dynamic elements that add a new type will raise an error!" #= none:27 =# notdsstr = "The solver types are not specialized in a static model context," * " which is sub-optimal." #= none:29 =# dynamic && (specialize && #= none:29 =# @warn(dsstr)) #= none:30 =# !dynamic && (!specialize && #= none:30 =# @info(notdsstr)) #= none:32 =# new(dynamic, iteration, printlevel, solutions, specialize, tabutime, tabulocal, tabudelta, threads, timelimit)

end`: DESCRIPTION

LocalSearchSolvers.Solver — Type

Solver <: AbstractSolver

Main solver. Handle the solving of a model, and optional multithreaded and/or distributed subsolvers.

Arguments:

model::Model: A formal description of the targeted problem

state::_State: An internal state to store the info necessary to a solving run

options::Options: User options for this solver

subs::Vector{_SubSolver}: Optional subsolvers

LocalSearchSolvers.Solver — Method

Solver{T}(m::Model; values::Dictionary{Int,T}=Dictionary{Int,T}()) where T <: Number
Solver{T}(;
    variables::Dictionary{Int,Variable}=Dictionary{Int,Variable}(),
    constraints::Dictionary{Int,Constraint}=Dictionary{Int,Constraint}(),
    objectives::Dictionary{Int,Objective}=Dictionary{Int,Objective}(),
    values::Dictionary{Int,T}=Dictionary{Int,T}(),
) where T <: Number

Constructor for a solver. Optional starting values can be provided.

# Model a sudoku model of size 4×4
m = sudoku(2)

# Create a solver instance with variables taking integral values
s = Solver{Int}(m)

# Solver with an empty model to be filled later and expected Float64 values
s = Solver{Float64}()

# Construct a solver from a sets of constraints, objectives, and variables.
s = Solver{Int}(
    variables = get_constraints(m),
    constraints = get_constraints(m),
    objectives = get_objectives(m)
)

Base.:∈ — Method

var::Int ∈ c::Constraint

Base.:∈ — Method

∈(x::Variable, constraint)
∈(value, x::Variable)

Check if a variable x is restricted by a constraint::Int, or if a value belongs to the domain of x.

LocalSearchSolvers.add! — Method

add!(m::M, x) where M <: Union{Model, AbstractSolver}
add!(m::M, c) where M <: Union{Model, AbstractSolver}
add!(m::M, o) where M <: Union{Model, AbstractSolver}

Add a variable x, a constraint c, or an objective o to m.

LocalSearchSolvers.add_value! — Method

add_value!(m::M, x, val) where M <: Union{Model, AbstractSolver}

Add val to x domain.

LocalSearchSolvers.add_var_to_cons! — Method

add_var_to_cons!(m::M, c, x) where M <: Union{Model, AbstractSolver}

Add x to the constraint c list of restricted variables.

LocalSearchSolvers.constraint! — Method

constraint!(m::M, func, vars) where M <: Union{Model, AbstractSolver}

Add a constraint with an error function func defined over variables vars.

LocalSearchSolvers.constraint — Method

constraint(f, vars)

DOCSTRING

LocalSearchSolvers.constriction — Method

constriction(m::M, x) where M <: Union{Model, AbstractSolver}

Return the constriction of variable x.

LocalSearchSolvers.delete_value! — Method

delete_value(m::M, x, val) where M <: Union{Model, AbstractSolver}

Delete val from x domain.

LocalSearchSolvers.delete_var_from_cons! — Method

delete_var_from_cons(m::M, c, x) where M <: Union{Model, AbstractSolver}

Delete x from the constraint c list of restricted variables.

LocalSearchSolvers.describe — Method

describe(m::M) where M <: Union{Model, AbstractSolver}

Describe the model.

LocalSearchSolvers.draw — Method

draw(m::M, x) where M <: Union{Model, AbstractSolver}

Draw a random value of x domain.

LocalSearchSolvers.get_cons_from_var — Method

get_cons_from_var(m::M, x) where M <: Union{Model, AbstractSolver}

Access the constraints restricting variable x.

LocalSearchSolvers.get_constraint — Method

get_constraint(m::M, c) where M <: Union{Model, AbstractSolver}

Access the constraint c.

LocalSearchSolvers.get_constraints — Method

get_constraints(m::M) where M <: Union{Model, AbstractSolver}

Access the constraints of m.

LocalSearchSolvers.get_domain — Method

get_domain(m::M, x) where M <: Union{Model, AbstractSolver}

Access the domain of variable x.

LocalSearchSolvers.get_name — Method

get_name(m::M, x) where M <: Union{Model, AbstractSolver}

Access the name of variable x.

LocalSearchSolvers.get_objective — Method

get_objective(m::M, o) where M <: Union{Model, AbstractSolver}

Access the objective o.

LocalSearchSolvers.get_objectives — Method

get_objectives(m::M) where M <: Union{Model, AbstractSolver}

Access the objectives of m.

LocalSearchSolvers.get_variable — Method

get_variable(m::M, x) where M <: Union{Model, AbstractSolver}

Access the variable x.

LocalSearchSolvers.get_variables — Method

get_variables(m::M) where M <: Union{Model, AbstractSolver}

Access the variables of m.

LocalSearchSolvers.get_vars_from_cons — Method

get_vars_from_cons(m::M, c) where M <: Union{Model, AbstractSolver}

Access the variables restricted by constraint c.

LocalSearchSolvers.golomb — Function

golomb(n, L=n²)

Model the Golomb problem of n marks on the ruler 0:L. The modeler argument accepts :raw, and :JuMP (default), which refer respectively to the solver internal model, the MathOptInterface model, and the JuMP model.

LocalSearchSolvers.length_cons — Method

length_cons(m::M, c) where M <: Union{Model, AbstractSolver}

Return the length of constraint c.

LocalSearchSolvers.length_cons — Method

length_cons(m::M) where M <: Union{Model, AbstractSolver}

Return the number of constraints in m.

LocalSearchSolvers.length_var — Method

length_var(m::M, x) where M <: Union{Model, AbstractSolver}

Return the domain length of variable x.

LocalSearchSolvers.mincut — Method

mincut(graph::AbstractMatrix{T}; source::Int, sink::Int, interdiction::Int = 0) where T <: Number

Compute the minimum cut of a graph.

Arguments:

graph: Any matrix <: AbstractMatrix that describes the capacities of the graph

source: Id of the source node; must be set

sink: Id of the sink node; must be set

interdiction: indicates the number of forbidden links

LocalSearchSolvers.model — Method

model()

Construct a _Model, empty by default. It is recommended to add the constraints, variables, and objectives from an empty _Model. The following keyword arguments are available,

vars=Dictionary{Int,Variable}(): collection of variables

cons=Dictionary{Int,Constraint}(): collection of cosntraints

objs=Dictionary{Int,Objective}(): collection of objectives

kind=:generic: the kind of problem modeled (useful for specialized methods such as pretty printing)

LocalSearchSolvers.o_dist_extrema — Method

dist_extrema(values::T...) where {T <: Number}

Computes the distance between extrema in an ordered set.

LocalSearchSolvers.o_mincut — Method

o_mincut(graph, values; interdiction = 0)

Compute the capacity of a cut (determined by the state of the solver) with a possible interdiction on the highest capacited links.

LocalSearchSolvers.objective! — Method

objective!(m::M, func) where M <: Union{Model, AbstractSolver}

Add an objective evaluated by func.

LocalSearchSolvers.objective — Method

objective(func, name)

Construct an objective with a function func that should be applied to a collection of variables.

LocalSearchSolvers.solution — Method

solution(s)

tabu Return the only/best known solution of a satisfaction/optimization model.

LocalSearchSolvers.solve! — Method

solve!(s; max_iteration=1000, verbose::Bool=false)

Run the solver until a solution is found or max_iteration is reached. verbose=true will print out details of the run.

# Simply run the solver with default max_iteration
solve!(s)

# Run indefinitely the solver with verbose behavior.
solve!(s, max_iteration = Inf, verbose = true)

LocalSearchSolvers.specialize! — Method

specialize!(s) = begin

DOCSTRING

LocalSearchSolvers.specialize — Method

specialize(m::M) where M <: Union{Model, AbstractSolver}

Specialize the structure of a model to avoid dynamic type attribution at runtime.

LocalSearchSolvers.sudoku — Method

sudoku(n; start= Dictionary{Int, Int}(), modeler = :JuMP)

Create a model for the sudoku problem of domain 1:n² with optional starting values. The modeler argument accepts :raw, :MOI, and :JuMP (default), which refer respectively to the solver internal model, the MathOptInterface model, and the JuMP model.

LocalSearchSolvers.variable! — Function

variable!(m::M, d) where M <: Union{Model, AbstractSolver}

Add a variable with domain d to m.

LocalSearchSolvers.variable — Method

variable(values::AbstractVector{T}, name::AbstractString; domain = :set) where T <: Number
variable(domain::AbstractDomain, name::AbstractString) where D <: AbstractDomain

Construct a variable with discrete domain. See the domain method for other options.

d = domain([1,2,3,4], types = :indices)
x1 = variable(d, "x1")
x2 = variable([-89,56,28], "x2", domain = :indices)