LocalSearchSolvers

LocalSearchSolvers.Solver

LocalSearchSolvers.all_different

LocalSearchSolvers.constraint

LocalSearchSolvers.describe

LocalSearchSolvers.domain

LocalSearchSolvers.is_sat

LocalSearchSolvers.is_specialized

LocalSearchSolvers.solve!

LocalSearchSolvers.specialize

LocalSearchSolvers.variable

LocalSearchSolvers.Solver — Method

Solver{T}(p::Problem; values::Dictionary{Int,T}=Dictionary{Int,T}()) where T <: Number
Solver{T}(;
    variables::Dictionary{Int,Variable}=Dictionary{Int,Variable}(),
    constraints::Dictionary{Int,Constraint}=Dictionary{Int,Constraint}(),
    objectives::Dictionary{Int,Objective}=Dictionary{Int,Objective}(),
    values::Dictionary{Int,T}=Dictionary{Int,T}(),
) where T <: Number

Constructor for a solver. Optional starting values can be provided.

# Model a sudoku problem of size 4×4
p = sudoku(2)

# Create a solver instance with variables taking integral values
s = Solver{Int}(p)

# Solver with an empty problem to be filled later and expected Float64 values
s = Solver{Float64}()

# Construct a solver from a sets of constraints, objectives, and variables.
s = Solver{Int}(
    variables = get_constraints(p),
    constraints = get_constraints(p),
    objectives = get_objectives(p)
)

LocalSearchSolvers.all_different — Method

all_different(x::Int...)

Global constraint ensuring that all the values of x are unique.

LocalSearchSolvers.constraint — Method

constraint(f::F, inds::Vector{Int}, values::Vector{T}) where {F <: Function,T <: Number}
constraint(f::F, inds::Vector{Int}, vars::Dictionary{Int,Variable}) where F <: Function

Test the validity of f over a set of values or draw them from a set of variables vars. Return a constraint if the test is succesful, otherwise raise an error.

LocalSearchSolvers.describe — Method

describe(p::Problem)
describe(s::AbstractSolver)

Describe the model of either a Problem or a Solver.

LocalSearchSolvers.domain — Method

domain(values::AbstractVector; type = :set)

Discrete domain constructor. The type keyword can be set to :set (default) or :indices.

d1 = domain([1,2,3,4], type = :indices)
d2 = domain([53.69, 89.2, 0.12])
d3 = domain([2//3, 89//123])

LocalSearchSolvers.is_sat — Method

is_sat(p::Problem)

Return true if p is a satisfaction problem.

LocalSearchSolvers.is_specialized — Method

is_specialized(p::Problem)
is_specialized(s::Solver)

Return true if the problem is already specialized.

LocalSearchSolvers.solve! — Method

solve!(s::Solver{T}; max_iteration=1000, verbose::Bool=false) where {T <: Real}

Run the solver until a solution is found or max_iteration is reached. verbose=true will print out details of the run.

# Simply run the solver with default max_iteration
solve!(s)

# Run indefinitely the solver with verbose behavior.
solve!(s, max_iteration = Inf, verbose = true)

LocalSearchSolvers.specialize — Method

specialize(p::Problem)
specialize(s::Solver)

Specialize the structure of a problem to avoid dynamic type attribution at runtime.

LocalSearchSolvers.variable — Method

variable(values::AbstractVector{T}, name::AbstractString; domain = :set) where T <: Number
variable(domain::AbstractDomain, name::AbstractString) where D <: AbstractDomain

Construct a variable with discrete domain. See the domain method for other options.

d = domain([1,2,3,4], types = :indices)
x1 = variable(d, "x1")
x2 = variable([-89,56,28], "x2", domain = :indices)