Internal

LocalSearchSolvers.AbstractSolver — Type

AbstractSolver

Abstract type to encapsulate the different solver types such as Solver or _SubSolver.

LocalSearchSolvers.Constraint — Type

Constraint{F <: Function}

Structure to store an error function and the variables it constrains.

LocalSearchSolvers.Objective — Type

Objective{F <: Function}

A structure to handle objectives in a solver. struct Objective{F <: Function} name::String f::F end`

LocalSearchSolvers.Objective — Method

Objective(F, o::Objective{F2}) where {F2 <: Function}

Constructor used in specializing a solver. Should never be called externally.

LocalSearchSolvers.SudokuInstance — Type

mutable struct SudokuInstance{T <: Integer} <: AbstractMatrix{T}

A struct for SudokuInstances, which is a subtype of AbstractMatrix.

SudokuInstance(A::AbstractMatrix{T})
SudokuInstance(::Type{T}, n::Int) # fill in blank sudoku of type T
SudokuInstance(n::Int) # fill in blank sudoku of type Int
SudokuInstance(::Type{T}) # fill in "standard" 9×9 sudoku of type T
SudokuInstance() # fill in "standard" 9×9 sudoku of type Int
SudokuInstance(n::Int, P::Pair{Tuple{Int, Int}, T}...) where {T <: Integer} # construct a sudoku given pairs of coordinates and values
SudokuInstance(P::Pair{Tuple{Int, Int}, T}...) # again, default to 9×9 sudoku, constructing given pairs

Constructor functions for the SudokuInstance struct.

LocalSearchSolvers.SudokuInstance — Method

SudokuInstance(X::Dictionary)

Construct a SudokuInstance with the values X of a solver as input.

LocalSearchSolvers.Variable — Type

Variable{D <: AbstractDomain}

A structure containing the necessary information for a solver's variables: name, domain, and constraints it belongs.

struct Variable{D <: AbstractDomain}
    name::String
    domain::D
    constraints::Indices{Int}
end

LocalSearchSolvers._State — Type

_State{T <: Number}

A mutable structure to store the internal state of a solver. All methods applied to _State are forwarded to S <: AbstractSolver.

mutable struct _State{T <: Number}
    values::Dictionary{Int,T} # TODO: handle multiple value type
    vars_costs::Dictionary{Int,Float64}
    cons_costs::Dictionary{Int,Float64}
    error::Float64
    tabu::Dictionary{Int,Int}
    optimizing::Bool
    best_solution::Dictionary{Int,T}
    best_solution_value::Union{Nothing,T}
end

LocalSearchSolvers._SubSolver — Type

_SubSolver <: AbstractSolver

An internal solver type called by Solver when multithreading is enabled.

Arguments:

id::Int: subsolver id for debugging

model::Model: a ref to the model of the main solver

state::_State: a deepcopy of the main solver that evolves independently

settings::Settings: a ref to the settings of the main solver

LocalSearchSolvers._SubSolver — Method

_SubSolver(ms::Solver, id)

Internal structure used in multithreading and distributed version of the solvers. It is only created at the start of a solve! run. Its behaviour regarding to sharing information is determined by the main Solver.

Base.Multimedia.display — Method

display(io::IO, S::SudokuInstance)
display(S::SudokuInstance) # default to stdout

Displays an $n\times n$ SudokuInstance.

Base.Multimedia.display — Method

Base.display(X::Dictionary)

Extends Base.display to a sudoku configuration.

Base.Multimedia.display — Method

Base.display(S::SudokuInstance)

Extends Base.display to SudokuInstance.

Base.getindex — Method

Base.getindex(S::SudokuInstance, i::Int)
Base.getindex(S::SudokuInstance, I::Vararg{Int,N}) where {N}

Extends Base.getindex for SudokuInstance.

Base.setindex! — Method

Base.setindex!(S::SudokuInstance, v, i::Int)
Base.setindex!(S::SudokuInstance, v, I::Vararg{Int,N})

Extends Base.setindex! for SudokuInstance.

Base.size — Method

Base.size(S::SudokuInstance)

Extends Base.size for SudokuInstance.

ConstraintDomains._add! — Method

_add!(c::Constraint, x)

Add the variable of indice x to c.

ConstraintDomains._delete! — Method

_delete!(c::Constraint, x::Int)

Delete x from c.

ConstraintDomains._length — Method

_length(c::Constraint)

Return the number of constrained variables by c.

LocalSearchSolvers._add_to_constraint! — Method

_add_to_constraint!(x::Variable, id)

Add a constraint id to the list of contraints of x.

LocalSearchSolvers._best! — Method

_best!(s::S, val, values = Dictionary()) where S <: Union{_State, AbstractSolver}

Set the best known value to val and, if values not empty, the best known solution.

LocalSearchSolvers._best — Method

_best(s::S) where S <: Union{_State, AbstractSolver}

Access the best known solution value (defined for optimization models only).

LocalSearchSolvers._check_restart — Method

_check_restart(s)

Check if a restart of s is necessary. If s has subsolvers, this check is independent for all of them.

LocalSearchSolvers._check_subs — Method

_check_subs(s)

Check if any subsolver of a main solver s, for

Satisfaction, has a solution, then return it, resume the run otherwise

Optimization, has a better solution, then assign it to its internal state

LocalSearchSolvers._compute! — Method

_compute!(s; o::Int = 1, cons_lst = Indices{Int}())

Compute the objective o's value if s is satisfied and return the current error.

Arguments:

s: a solver

o: targeted objective

cons_lst: list of targeted constraints, if empty compute for the whole set

LocalSearchSolvers._compute_cost! — Method

_compute_cost!(s, ind, c)

Compute the cost of constraint c with index ind.

LocalSearchSolvers._compute_costs! — Method

_compute_costs!(s; cons_lst::Indices{Int} = Indices{Int}())

Compute the cost of constraints c in cons_lst. If cons_lst is empty, compute the cost for all the constraints in s.

LocalSearchSolvers._compute_objective! — Method

_compute_objective!(s, o::Objective)
_compute_objective!(s, o = 1)

Compute the objective o's value.

LocalSearchSolvers._cons_cost! — Method

_cons_cost!(s::S, c, cost) where S <: Union{_State, AbstractSolver}

Set the cost of constraint c.

LocalSearchSolvers._cons_cost — Method

_cons_cost(s::S, c) where S <: Union{_State, AbstractSolver}

Return the cost of constraint c.

LocalSearchSolvers._cons_costs! — Method

_cons_costs!(s::S, costs) where S <: Union{_State, AbstractSolver}

Set the constraints costs.

LocalSearchSolvers._cons_costs — Method

_cons_costs(s::S) where S <: Union{_State, AbstractSolver}

Access the constraints costs.

LocalSearchSolvers._constriction — Method

_constriction(x::Variable)

Return the cosntriction of x, i.e. the number of constraints restricting x.

LocalSearchSolvers._decay_tabu! — Method

_decay_tabu!(s::S) where S <: Union{_State, AbstractSolver}

Decay the tabu list.

LocalSearchSolvers._decrease_tabu! — Method

_decrease_tabu!(s::S, x) where S <: Union{_State, AbstractSolver}

Decrement the tabu value of variable x.

LocalSearchSolvers._delete_from_constraint! — Method

_delete_from_constraint!(x::Variable, id)

Delete a constraint id from the list of contraints of x.

LocalSearchSolvers._delete_tabu! — Method

_delete_tabu!(s::S, x) where S <: Union{_State, AbstractSolver}

Delete the tabu entry of variable x.

LocalSearchSolvers._draw! — Method

_draw!(s)

Draw a random (re-)starting configuration.

LocalSearchSolvers._empty_tabu! — Method

_empty_tabu!(s::S) where S <: Union{_State, AbstractSolver}

Empty the tabu list.

LocalSearchSolvers._error! — Method

_error!(s::S, val) where S <: Union{_State, AbstractSolver}

Set the error of the current state of s to val.

LocalSearchSolvers._error — Method

_error(s::S) where S <: Union{_State, AbstractSolver}

Access the error of the current state of s.

LocalSearchSolvers._find_rand_argmax — Method

_find_rand_argmax(d::DictionaryView)

Compute argmax of d and select one element randomly.

LocalSearchSolvers._format_line — Method

_format_line(r, M)

Format line of a sudoku grid.

LocalSearchSolvers._format_line_segment — Method

_format_line_segment(r, col_pos, M)

Format line segment of a sudoku grid.

LocalSearchSolvers._format_val — Method

_format_val(a)

Format an integer a into a string for SudokuInstance.

LocalSearchSolvers._get_constraints — Method

_get_constraints(x::Variable)

Access the list of constraints of x.

LocalSearchSolvers._get_name — Method

_get_name(x::Variable)

Access the name of x.

LocalSearchSolvers._get_sep_line — Method

_get_sep_line(s, pos_row, M)

Return a line separator.

LocalSearchSolvers._get_vars — Method

_get_vars(c::Constraint)

Returns the variables constrained by c.

LocalSearchSolvers._inc_cons! — Method

_inc_vars!(m::M) where M <: Union{Model, AbstractSolver}

Increment the maximum constraint id that has been attributed to m.

LocalSearchSolvers._inc_objs! — Method

_inc_vars!(m::M) where M <: Union{Model, AbstractSolver}

Increment the maximum objective id that has been attributed to m.

LocalSearchSolvers._inc_vars! — Method

_inc_vars!(m::M) where M <: Union{Model, AbstractSolver}

Increment the maximum variable id that has been attributed to m.

LocalSearchSolvers._init_solve! — Method

_init_solve!(s::S) where S <: AbstractSolver

Initialize a solver in both sequential and parallel contexts.

LocalSearchSolvers._insert_tabu! — Method

_insert_tabu!(s::S, x, tabu_time) where S <: Union{_State, AbstractSolver}

Insert the bariable x as tabu for tabu_time.

LocalSearchSolvers._length_tabu — Method

_length_tabu!(s::S) where S <: Union{_State, AbstractSolver}

Return the length of the tabu list.

LocalSearchSolvers._make_settings! — Method

_make_settings!(settings)

Make default settings unless, for each setting, already provided by the user.

LocalSearchSolvers._max_cons — Method

_max_cons(m::M) where M <: Union{Model, AbstractSolver}

Access the maximum constraint id that has been attributed to m.

LocalSearchSolvers._max_objs — Method

_max_objs(m::M) where M <: Union{Model, AbstractSolver}

Access the maximum objective id that has been attributed to m.

LocalSearchSolvers._max_vars — Method

_max_vars(m::M) where M <: Union{Model, AbstractSolver}

Access the maximum variable id that has been attributed to m.

LocalSearchSolvers._move! — Function

_move!(s, x::Int, dim::Int = 0)

Perform an improving move in x neighbourhood if possible.

Arguments:

s: a solver of type S <: AbstractSolver

x: selected variable id

dim: describe the dimension of the considered neighbourhood

LocalSearchSolvers._optimizing! — Method

_optimizing!(s::S) where S <: Union{_State, AbstractSolver}

Set the solver optimizing status to true.

LocalSearchSolvers._optimizing — Method

_optimizing(s::S) where S <: Union{_State, AbstractSolver}

Check if s is in an optimizing state.

LocalSearchSolvers._restart! — Function

_restart!(s, k = 10)

Restart a solver.

LocalSearchSolvers._satisfying! — Method

_satisfying!(s::S) where S <: Union{_State, AbstractSolver}

Set the solver optimizing status to false.

LocalSearchSolvers._select_worse — Method

_select_worse(s::S) where S <: Union{_State, AbstractSolver}

Within the non-tabu variables, select the one with the worse error .

LocalSearchSolvers._set! — Method

_set!(s::S, x, val) where S <: Union{_State, AbstractSolver}

Set the value of variable x to val.

LocalSearchSolvers._solution! — Method

_solution!(s::S, values) where S <: Union{_State, AbstractSolver}

Set the best known solution to values.

LocalSearchSolvers._solution — Method

_solution(s::S) where S <: Union{_State, AbstractSolver}

Access the best known solution.

LocalSearchSolvers._solve! — Method

_solve!(s, stop)

Start a solving run of a subsolver with a shared Atomic boolean stop.

LocalSearchSolvers._step! — Method

_step!(s)

Iterate a step of the solver run.

LocalSearchSolvers._swap_value! — Method

_set!(s::S, x, y) where S <: Union{_State, AbstractSolver}

Swap the values of variables x and y.

LocalSearchSolvers._tabu! — Method

_tabu!(s::S, tabu) where S <: Union{_State, AbstractSolver}

Set the variables tabu list.

LocalSearchSolvers._tabu — Method

_tabu(s::S, x) where S <: Union{_State, AbstractSolver}

Return the tabu value of variable x.

LocalSearchSolvers._tabu — Method

_tabu(s::S) where S <: Union{_State, AbstractSolver}

Access the list of tabu variables.

LocalSearchSolvers._to_union — Method

_to_union(datatype)

Make a minimal Union type from a collection of data types.

LocalSearchSolvers._value! — Method

_value!(s::S, x, val) where S <: Union{_State, AbstractSolver}

Set the value of variable x to val.

LocalSearchSolvers._value — Method

_value(s::S, x) where S <: Union{_State, AbstractSolver}

Return the value of variable x.

LocalSearchSolvers._values! — Method

_values!(s::S, values) where S <: Union{_State, AbstractSolver}

Set the variables values.

LocalSearchSolvers._values — Method

_vars_costs(s::S) where S <: Union{_State, AbstractSolver}

Access the variables costs.

LocalSearchSolvers._var_cost! — Method

_var_cost!(s::S, x, cost) where S <: Union{_State, AbstractSolver}

Set the cost of variable x.

LocalSearchSolvers._var_cost — Method

_var_cost(s::S, x) where S <: Union{_State, AbstractSolver}

Return the cost of variable x.

LocalSearchSolvers._vars_costs! — Method

_vars_costs!(s::S, costs) where S <: Union{_State, AbstractSolver}

Set the variables costs.

LocalSearchSolvers._vars_costs — Method

_vars_costs(s::S) where S <: Union{_State, AbstractSolver}

Access the variables costs.

LocalSearchSolvers._verbose — Method

_verbose(settings, str)

Temporary logging function. #TODO: use better log instead (LoggingExtra.jl)

LocalSearchSolvers.get_kind — Method

get_kind(m::M) where M <: Union{Model, AbstractSolver}

Access the kind of m, such as :sudoku or :generic (default).

LocalSearchSolvers.is_sat — Method

is_sat(m::M) where M <: Union{Model, AbstractSolver}

Return true if m is a satisfaction model.

LocalSearchSolvers.is_specialized — Method

is_specialized(m::M) where M <: Union{Model, AbstractSolver}

Return true if the model is already specialized.

LocalSearchSolvers.length_objs — Method

length_objs(m::M) where M <: Union{Model, AbstractSolver}

Return the number of objectives in m.

LocalSearchSolvers.length_vars — Method

length_vars(m::M) where M <: Union{Model, AbstractSolver}

Return the number of variables in m.

LocalSearchSolvers.setting — Method

setting(s, sym::Symbol)

Retrieve a specific setting sym of solver s.