Internal

LocalSearchSolvers.AbstractSolver — Type

AbstractSolver

Abstract type to encapsulate the different solver types such as Solver or _SubSolver.

LocalSearchSolvers.Constraint — Type

Constraint{F <: Function}

Structure to store an error function and the variables it constrains.

LocalSearchSolvers.LeadSolver — Type

LeadSolver <: MetaSolver

Solver managed remotely by a MainSolver. Can manage its own set of local sub solvers.

LocalSearchSolvers.MainSolver — Type

MainSolver <: AbstractSolver

Main solver. Handle the solving of a model, and optional multithreaded and/or distributed subsolvers.

Arguments:

model::Model: A formal description of the targeted problem

state::_State: An internal state to store the info necessary to a solving run

options::Options: User options for this solver

subs::Vector{_SubSolver}: Optional subsolvers

LocalSearchSolvers.MetaSolver — Type

Abstract type to encapsulate all solver types that manages other solvers.

LocalSearchSolvers.Objective — Type

Objective{F <: Function}

A structure to handle objectives in a solver. struct Objective{F <: Function} name::String f::F end`

LocalSearchSolvers.Objective — Method

Objective(F, o::Objective{F2}) where {F2 <: Function}

Constructor used in specializing a solver. Should never be called externally.

LocalSearchSolvers.Variable — Type

Variable{D <: AbstractDomain}

A structure containing the necessary information for a solver's variables: name, domain, and constraints it belongs.

struct Variable{D <: AbstractDomain}
    domain::D
    constraints::Indices{Int}
end

LocalSearchSolvers._Model — Type

_Model{V <: Variable{<:AbstractDomain},C <: Constraint{<:Function},O <: Objective{<:Function}}

A struct to model a problem as a set of variables, domains, constraints, and objectives.

struct _Model{V <: Variable{<:AbstractDomain},C <: Constraint{<:Function},O <: Objective{<:Function}}
    variables::Dictionary{Int,V}
    constraints::Dictionary{Int,C}
    objectives::Dictionary{Int,O}

    # counter to add new variables: vars, cons, objs
    max_vars::Ref{Int}
    max_cons::Ref{Int}
    max_objs::Ref{Int}

    # Bool to indicate if the _Model instance has been specialized (relatively to types)
    specialized::Ref{Bool}

    # Symbol to indicate the kind of model for specialized methods such as pretty printing
    kind::Symbol
end

LocalSearchSolvers._State — Type

GeneralState{T <: Number}

A mutable structure to store the general state of a solver. All methods applied to GeneralState are forwarded to S <: AbstractSolver.

mutable struct GeneralState{T <: Number} <: AbstractState
    configuration::Configuration{T}
    cons_costs::Dictionary{Int, Float64}
    last_improvement::Int
    tabu::Dictionary{Int, Int}
    vars_costs::Dictionary{Int, Float64}
end

LocalSearchSolvers._SubSolver — Type

_SubSolver <: AbstractSolver

An internal solver type called by MetaSolver when multithreading is enabled.

Arguments:

id::Int: subsolver id for debugging

model::Model: a ref to the model of the main solver

state::_State: a deepcopy of the main solver that evolves independently

options::Options: a ref to the options of the main solver

Base.empty! — Method

empty!(s::Solver)

Base.empty! — Method

empty!(m::Model)

DOCSTRING

Base.in — Method

var::Int ∈ c::Constraint

Base.in — Method

x::Variable ∈ constraint
value ∈ x::Variable

Check if a variable x is restricted by a constraint::Int, or if a value belongs to the domain of x.

LocalSearchSolvers._add! — Method

_add!(c::Constraint, x)

Add the variable of indice x to c.

LocalSearchSolvers._add_to_constraint! — Method

_add_to_constraint!(x::Variable, id)

Add a constraint id to the list of contraints of x.

LocalSearchSolvers._check_restart — Method

_check_restart(s)

Check if a restart of s is necessary. If s has subsolvers, this check is independent for all of them.

LocalSearchSolvers._check_subs — Method

_check_subs(s)

Check if any subsolver of a main solver s, for

Satisfaction, has a solution, then return it, resume the run otherwise

Optimization, has a better solution, then assign it to its internal state

LocalSearchSolvers._compute! — Method

_compute!(s; o::Int = 1, cons_lst = Indices{Int}())

Compute the objective o's value if s is satisfied and return the current error.

Arguments:

s: a solver

o: targeted objective

cons_lst: list of targeted constraints, if empty compute for the whole set

LocalSearchSolvers._compute_cost! — Method

_compute_cost!(s, ind, c)

Compute the cost of constraint c with index ind.

LocalSearchSolvers._compute_costs! — Method

_compute_costs!(s; cons_lst::Indices{Int} = Indices{Int}())

Compute the cost of constraints c in cons_lst. If cons_lst is empty, compute the cost for all the constraints in s.

LocalSearchSolvers._compute_objective! — Method

_compute_objective!(s, o::Objective)
_compute_objective!(s, o = 1)

Compute the objective o's value.

LocalSearchSolvers._cons_cost! — Method

_cons_cost!(s::S, c, cost) where S <: Union{_State, AbstractSolver}

Set the cost of constraint c.

LocalSearchSolvers._cons_cost — Method

_cons_cost(s::S, c) where S <: Union{_State, AbstractSolver}

Return the cost of constraint c.

LocalSearchSolvers._cons_costs! — Method

_cons_costs!(s::S, costs) where S <: Union{_State, AbstractSolver}

Set the constraints costs.

LocalSearchSolvers._cons_costs — Method

_cons_costs(s::S) where S <: Union{_State, AbstractSolver}

Access the constraints costs.

LocalSearchSolvers._constriction — Method

_constriction(x::Variable)

Return the cosntriction of x, i.e. the number of constraints restricting x.

LocalSearchSolvers._delete! — Method

_delete!(c::Constraint, x::Int)

Delete x from c.

LocalSearchSolvers._delete_from_constraint! — Method

_delete_from_constraint!(x::Variable, id)

Delete a constraint id from the list of contraints of x.

LocalSearchSolvers._draw! — Method

_draw!(s)

Draw a random (re-)starting configuration.

LocalSearchSolvers._find_rand_argmax — Method

_find_rand_argmax(d::DictionaryView)

Compute argmax of d and select one element randomly.

LocalSearchSolvers._get_constraints — Method

_get_constraints(x::Variable)

Access the list of constraints of x.

LocalSearchSolvers._get_vars — Method

_get_vars(c::Constraint)

Returns the variables constrained by c.

LocalSearchSolvers._inc_cons! — Method

_inc_vars!(m::M) where M <: Union{Model, AbstractSolver}

Increment the maximum constraint id that has been attributed to m.

LocalSearchSolvers._inc_objs! — Method

_inc_vars!(m::M) where M <: Union{Model, AbstractSolver}

Increment the maximum objective id that has been attributed to m.

LocalSearchSolvers._inc_vars! — Method

_inc_vars!(m::M) where M <: Union{Model, AbstractSolver}

Increment the maximum variable id that has been attributed to m.

LocalSearchSolvers._is_empty — Method

_is_empty(m::Model)

DOCSTRING

LocalSearchSolvers._length — Method

_length(c::Constraint)

Return the number of constrained variables by c.

LocalSearchSolvers._max_cons — Method

_max_cons(m::M) where M <: Union{Model, AbstractSolver}

Access the maximum constraint id that has been attributed to m.

LocalSearchSolvers._max_objs — Method

_max_objs(m::M) where M <: Union{Model, AbstractSolver}

Access the maximum objective id that has been attributed to m.

LocalSearchSolvers._max_vars — Method

_max_vars(m::M) where M <: Union{Model, AbstractSolver}

Access the maximum variable id that has been attributed to m.

LocalSearchSolvers._move! — Function

_move!(s, x::Int, dim::Int = 0)

Perform an improving move in x neighbourhood if possible.

Arguments:

s: a solver of type S <: AbstractSolver

x: selected variable id

dim: describe the dimension of the considered neighbourhood

LocalSearchSolvers._neighbours — Function

_neighbours(s, x, dim = 0)

DOCSTRING

Arguments:

s: DESCRIPTION

x: DESCRIPTION

dim: DESCRIPTION

LocalSearchSolvers._optimizing! — Method

_optimizing!(s::S) where S <: Union{_State, AbstractSolver}

Set the solver optimizing status to true.

LocalSearchSolvers._optimizing — Method

_optimizing(s::S) where S <: Union{_State, AbstractSolver}

Check if s is in an optimizing state.

LocalSearchSolvers._print_level! — Method

_print_level!(options, level) = begin

DOCSTRING

LocalSearchSolvers._restart! — Function

_restart!(s, k = 10)

Restart a solver.

LocalSearchSolvers._satisfying! — Method

_satisfying!(s::S) where S <: Union{_State, AbstractSolver}

Set the solver optimizing status to false.

LocalSearchSolvers._select_worse — Method

_select_worse(s::S) where S <: Union{_State, AbstractSolver}

Within the non-tabu variables, select the one with the worse error .

LocalSearchSolvers._set! — Method

_set!(s::S, x, val) where S <: Union{_State, AbstractSolver}

Set the value of variable x to val.

LocalSearchSolvers._specialize — Method

_specialize(options) = begin

DOCSTRING

LocalSearchSolvers._step! — Method

_step!(s)

Iterate a step of the solver run.

LocalSearchSolvers._swap_value! — Method

_set!(s::S, x, y) where S <: Union{_State, AbstractSolver}

Swap the values of variables x and y.

LocalSearchSolvers._tabu_delta! — Method

_tabu_delta!(options, time) = begin

DOCSTRING

LocalSearchSolvers._tabu_delta — Method

_tabu_delta(options) = begin

DOCSTRING

LocalSearchSolvers._tabu_local! — Method

_tabu_local!(options, time) = begin

DOCSTRING

LocalSearchSolvers._tabu_local — Method

_tabu_local(options) = begin

DOCSTRING

LocalSearchSolvers._tabu_time! — Method

_tabu_time!(options, time) = begin

DOCSTRING

LocalSearchSolvers._tabu_time — Method

_tabu_time(options) = begin

DOCSTRING

LocalSearchSolvers._threads! — Method

_threads!(options, threads) = begin

DOCSTRING

LocalSearchSolvers._threads — Method

_threads(options) = begin

DOCSTRING

LocalSearchSolvers._to_union — Method

_to_union(datatype)

Make a minimal Union type from a collection of data types.

LocalSearchSolvers._value! — Method

_value!(s::S, x, val) where S <: Union{_State, AbstractSolver}

Set the value of variable x to val.

LocalSearchSolvers._value — Method

_value(s::S, x) where S <: Union{_State, AbstractSolver}

Return the value of variable x.

LocalSearchSolvers._values! — Method

_values!(s::S, values) where S <: Union{_State, AbstractSolver}

Set the variables values.

LocalSearchSolvers._values — Method

_vars_costs(s::S) where S <: Union{_State, AbstractSolver}

Access the variables costs.

LocalSearchSolvers._var_cost! — Method

_var_cost!(s::S, x, cost) where S <: Union{_State, AbstractSolver}

Set the cost of variable x.

LocalSearchSolvers._var_cost — Method

_var_cost(s::S, x) where S <: Union{_State, AbstractSolver}

Return the cost of variable x.

LocalSearchSolvers._vars_costs! — Method

_vars_costs!(s::S, costs) where S <: Union{_State, AbstractSolver}

Set the variables costs.

LocalSearchSolvers._vars_costs — Method

_vars_costs(s::S) where S <: Union{_State, AbstractSolver}

Access the variables costs.

LocalSearchSolvers._verbose — Method

_verbose(settings, str)

Temporary logging function. #TODO: use better log instead (LoggingExtra.jl)

LocalSearchSolvers.decay_tabu! — Method

_decay_tabu!(s::S) where S <: Union{_State, AbstractSolver}

Decay the tabu list.

LocalSearchSolvers.decrease_tabu! — Method

_decrease_tabu!(s::S, x) where S <: Union{_State, AbstractSolver}

Decrement the tabu value of variable x.

LocalSearchSolvers.delete_tabu! — Method

_delete_tabu!(s::S, x) where S <: Union{_State, AbstractSolver}

Delete the tabu entry of variable x.

LocalSearchSolvers.domain_size — Method

domain_size(m::Model, x) = begin

DOCSTRING

LocalSearchSolvers.empty_tabu! — Method

_empty_tabu!(s::S) where S <: Union{_State, AbstractSolver}

Empty the tabu list.

LocalSearchSolvers.get_kind — Method

get_kind(m::M) where M <: Union{Model, AbstractSolver}

Access the kind of m, such as :sudoku or :generic (default).

LocalSearchSolvers.get_time_stamp — Method

get_time_stamp(m::M) where M <: Union{Model, AbstractSolver}

Access the time (since epoch) when the model was created. This time stamp is for internal performance measurement.

LocalSearchSolvers.insert_tabu! — Method

_insert_tabu!(s::S, x, tabu_time) where S <: Union{_State, AbstractSolver}

Insert the bariable x as tabu for tabu_time.

LocalSearchSolvers.is_specialized — Method

is_specialized(m::M) where M <: Union{Model, AbstractSolver}

Return true if the model is already specialized.

LocalSearchSolvers.length_objs — Method

length_objs(m::M) where M <: Union{Model, AbstractSolver}

Return the number of objectives in m.

LocalSearchSolvers.length_tabu — Method

_length_tabu!(s::S) where S <: Union{_State, AbstractSolver}

Return the length of the tabu list.

LocalSearchSolvers.length_vars — Method

length_vars(m::M) where M <: Union{Model, AbstractSolver}

Return the number of variables in m.

LocalSearchSolvers.post_process — Method

post_process(s::MainSolver)

Launch a serie of tasks to round-up a solving run, for instance, export a run's info.

LocalSearchSolvers.remote_dispatch! — Method

remote_dispatch!(solver)

Starts the LeadSolvers attached to the MainSolver.

LocalSearchSolvers.remote_stop! — Method

remote_stop!!(solver)

Fetch the pool of solutions from LeadSolvers and merge it into the MainSolver.

LocalSearchSolvers.solve_for_loop! — Method

solve_for_loop!(solver, stop, sat, iter)

First loop in the solving process that starts LeadSolvers from the MainSolver, and _SubSolvers from each MetaSolver.

LocalSearchSolvers.solve_while_loop! — Method

solve_while_loop!(s, )

Search the space of configurations.

LocalSearchSolvers.stop_while_loop — Method

stop_while_loop()

Check the stop conditions of the solve! while inner loop.

LocalSearchSolvers.tabu_list — Method

_tabu(s::S) where S <: Union{_State, AbstractSolver}

Access the list of tabu variables.

LocalSearchSolvers.tabu_value — Method

_tabu(s::S, x) where S <: Union{_State, AbstractSolver}

Return the tabu value of variable x.