LocalSearchSolvers.Solver

LocalSearchSolvers.c_all_different

LocalSearchSolvers.c_all_equal

LocalSearchSolvers.c_dist_different

LocalSearchSolvers.c_ordered

LocalSearchSolvers.constraint

LocalSearchSolvers.describe

LocalSearchSolvers.domain

LocalSearchSolvers.is_sat

LocalSearchSolvers.is_specialized

LocalSearchSolvers.o_dist_extrema

LocalSearchSolvers.solve!

LocalSearchSolvers.specialize

LocalSearchSolvers.variable

LocalSearchSolvers.Solver — Method

Solver{T}(m::Model; values::Dictionary{Int,T}=Dictionary{Int,T}()) where T <: Number
Solver{T}(;
    variables::Dictionary{Int,Variable}=Dictionary{Int,Variable}(),
    constraints::Dictionary{Int,Constraint}=Dictionary{Int,Constraint}(),
    objectives::Dictionary{Int,Objective}=Dictionary{Int,Objective}(),
    values::Dictionary{Int,T}=Dictionary{Int,T}(),
) where T <: Number

Constructor for a solver. Optional starting values can be provided.

# Model a sudoku model of size 4×4
m = sudoku(2)

# Create a solver instance with variables taking integral values
s = Solver{Int}(m)

# Solver with an empty model to be filled later and expected Float64 values
s = Solver{Float64}()

# Construct a solver from a sets of constraints, objectives, and variables.
s = Solver{Int}(
    variables = get_constraints(m),
    constraints = get_constraints(m),
    objectives = get_objectives(m)
)

LocalSearchSolvers.c_all_different — Method

all_different(x::Int...)

Global constraint ensuring that all the values of x are unique.

LocalSearchSolvers.c_all_equal — Method

all_equal(x::Int...; param::T)
all_equal(x::Int...)

Global constraint ensuring that all the values of x are all_equal (to param if given).

LocalSearchSolvers.c_dist_different — Method

dist_different(i::T, j::T, k::T, l::T) where {T <: Number}

Local constraint ensuring that |i - j| ≠ |k - l|.

LocalSearchSolvers.c_ordered — Method

ordered(x::Int...)

Global constraint ensuring that all the values of x are ordered.

LocalSearchSolvers.constraint — Method

constraint(f::F, inds::Vector{Int}, values::Vector{T}) where {F <: Function,T <: Number}
constraint(f::F, inds::Vector{Int}, vars::Dictionary{Int,Variable}) where F <: Function

Test the validity of f over a set of values or draw them from a set of variables vars. Return a constraint if the test is succesful, otherwise raise an error.

LocalSearchSolvers.describe — Method

describe(m::Model)
describe(s::AbstractSolver)

Describe the model of either a Model or a Solver.

LocalSearchSolvers.domain — Method

domain(values::AbstractVector; type = :set)

Discrete domain constructor. The type keyword can be set to :set (default) or :indices.

d1 = domain([1,2,3,4], type = :indices)
d2 = domain([53.69, 89.2, 0.12])
d3 = domain([2//3, 89//123])

LocalSearchSolvers.is_sat — Method

is_sat(m::Model)

Return true if p is a satisfaction model.

LocalSearchSolvers.is_specialized — Method

is_specialized(m::Model)
is_specialized(s::Solver)

Return true if the model is already specialized.

LocalSearchSolvers.o_dist_extrema — Method

dist_extrema(values::T...) where {T <: Number}

Computes the distance between extrema in an ordered set.

LocalSearchSolvers.solve! — Method

solve!(s::Solver{T}; max_iteration=1000, verbose::Bool=false) where {T <: Real}

Run the solver until a solution is found or max_iteration is reached. verbose=true will print out details of the run.

# Simply run the solver with default max_iteration
solve!(s)

# Run indefinitely the solver with verbose behavior.
solve!(s, max_iteration = Inf, verbose = true)

LocalSearchSolvers.specialize — Method

specialize(m::Model)
specialize(s::Solver)

Specialize the structure of a model to avoid dynamic type attribution at runtime.

LocalSearchSolvers.variable — Method

variable(values::AbstractVector{T}, name::AbstractString; domain = :set) where T <: Number
variable(domain::AbstractDomain, name::AbstractString) where D <: AbstractDomain

Construct a variable with discrete domain. See the domain method for other options.

d = domain([1,2,3,4], types = :indices)
x1 = variable(d, "x1")
x2 = variable([-89,56,28], "x2", domain = :indices)